Not necessary to use uint64_t for number of blocks owned by GC, size_t suffices.  Micro-optimization.  (r+me)

+++ MMgc/GC-inlines.h
@@ -488,7 +488,7 @@ namespace MMgc
 
 	REALLY_INLINE size_t GC::GetNumBlocks()
 	{
-		return (size_t)policy.blocksOwnedByGC();
+		return policy.blocksOwnedByGC();
 	}
 
 	REALLY_INLINE void* GC::allocaTop()
